move logic to base class
